+++ .travis.yml
@@ -33,7 +33,7 @@ cache:
 env:
   global:
     # If changing this number, please also change it in `testing/testing.go`.
-    - STRIPE_MOCK_VERSION=0.60.0
+    - STRIPE_MOCK_VERSION=0.63.0
 
 go:
   - "1.9.x"
