+++ balance_transaction_test.go
